folder restructure cleanup
