+++ .gitignore
@@ -6,6 +6,7 @@
 
 # Ignore bundler config.
 /client/bundle.js
+/public/bundle.js
 bundle.js
 .vscode/
 /node_modules
