+++ client/index.js
@@ -1,26 +1,10 @@
 import React from "react"
-import ReactDOM from "react-dom"
-import { HashRouter, Route, Switch } from "react-router-dom"
-import { createStore, compose, applyMiddleware } from "redux"
-import createSagaMiddleware from "redux-saga"
-import reducer from "./reducers/root_reducer"
-import rootSagaWatcher from "./sagas"
-import { Provider } from "react-redux"
-import logger from "redux-logger"
-import App from "./components/App"
+import reducers from "./reducers/root_reducer"
+import sagas from "./sagas"
+import { App } from "./App"
+import { configureStore } from "./configureStore"
+import { render } from "react-dom"
 
-const sagaMiddleware = createSagaMiddleware()
-const middlewares = applyMiddleware(sagaMiddleware, logger)
-const store = createStore(reducer, compose(middlewares))
-sagaMiddleware.run(rootSagaWatcher)
+const { store } = configureStore(reducers, sagas)
 
-ReactDOM.render(
-  <Provider store={store}>
-    <HashRouter>
-      <Switch>
-        <Route exact path="/" component={App} />
-      </Switch>
-    </HashRouter>
-  </Provider>,
-  document.getElementById("root")
-)
+render(<App store={store} />, document.getElementById("root"))
