+++ client/lib/styles/spaces.js
@@ -1,6 +1,4 @@
-const spaces = {
+export const spaces = {
   sm: "8px",
   md: "16px"
 }
-
-export default spaces
