+++ package.json
@@ -46,6 +46,7 @@
     "babel-loader": "^8.0.4",
     "cors": "^2.8.5",
     "enzyme": "^3.8.0",
+    "enzyme-adapter-react-16": "^1.7.1",
     "eslint": "4.x",
     "eslint-config-prettier": "^3.3.0",
     "eslint-plugin-prettier": "^3.0.0",
