+++ webpack.config.js
@@ -10,7 +10,7 @@ module.exports = {
   mode: config.mode,
   entry: ["babel-polyfill", "./client"],
   output: {
-    path: path.join(__dirname, "client/resources"),
+    path: path.join(__dirname, "public"),
     filename: "bundle.js"
   },
   module: {
